Updated UI for every fragment, implemented comments, and uploaded walkthrough GIF
