+++ app/src/main/res/layout/fragment_detail.xml
@@ -1,14 +1,116 @@
 <?xml version="1.0" encoding="utf-8"?>
-<FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
+
+<ScrollView
+    xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context=".Fragments.DetailFragment">
+    android:layout_height="match_parent">
 
-    <!-- TODO: Update blank fragment layout -->
-    <TextView
+    <RelativeLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        android:text="@string/hello_blank_fragment" />
+        android:layout_marginTop="10dp"
+        tools:context=".fragments.PostDetailsFragment">
 
-</FrameLayout>
+        <TextView
+            android:id="@+id/tvPostDescription"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:layout_alignParentEnd="true"
+            android:layout_marginStart="12dp"
+            android:layout_marginEnd="15dp"
+            android:layout_toRightOf="@id/ivPosterProfile"
+            android:minLines="1"
+            tools:text="This is a comment" />
+
+        <TextView
+            android:id="@+id/tvTimestamp"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:layout_below="@id/tvPostDescription"
+            android:layout_marginStart="12dp"
+            android:layout_marginTop="5dp"
+            android:layout_toRightOf="@id/ivPosterProfile"
+            android:textSize="11sp"
+            tools:text="1 day ago" />
+
+        <ImageView
+            android:id="@+id/ivPosterProfile"
+            android:layout_width="40dp"
+            android:layout_height="40dp"
+            android:layout_alignParentStart="true"
+            android:layout_marginStart="15dp"
+            tools:src="@tools:sample/avatars" />
+
+        <View
+            android:id="@+id/divider"
+            android:layout_width="match_parent"
+            android:layout_height="0.5dp"
+            android:layout_below="@id/tvTimestamp"
+            android:layout_marginTop="10dp"
+            android:background="@android:color/darker_gray" />
+
+        <androidx.recyclerview.widget.RecyclerView
+            android:id="@+id/rvComments"
+            android:layout_width="match_parent"
+            android:layout_height="match_parent"
+            android:layout_below="@id/divider"/>
+
+        <View
+            android:id="@+id/otherdivider"
+            android:layout_width="match_parent"
+            android:layout_height="0.5dp"
+            android:layout_below="@id/rvComments"
+            android:layout_marginTop="10dp"
+            android:background="@android:color/darker_gray"
+            android:layout_above="@+id/rlNewComment"/>
+
+        <RelativeLayout
+            android:id="@+id/rlNewComment"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:layout_below="@id/rvComments"
+            android:layout_marginStart="15dp"
+            android:layout_marginTop="10dp"
+            android:layout_marginEnd="15dp">
+
+            <EditText
+                android:id="@+id/etComment"
+                android:layout_width="230dp"
+                android:layout_height="match_parent"
+                android:layout_centerVertical="true"
+                android:layout_marginStart="20dp"
+                android:layout_toRightOf="@id/ivCurrUserProfile"
+                android:background="@android:color/transparent"
+                android:ems="10"
+                android:hint="Add a comment..."
+                android:inputType="textPersonName"
+                android:minLines="1"
+                android:textSize="14sp" />
+
+            <ImageView
+                android:id="@+id/ivCurrUserProfile"
+                android:layout_width="40dp"
+                android:layout_height="40dp"
+                app:srcCompat="@drawable/instagram_user_filled_24" />
+
+            <Button
+                android:id="@+id/btnComment"
+                android:layout_width="10dp"
+                android:layout_height="35dp"
+                android:layout_alignParentEnd="true"
+                android:layout_centerVertical="true"
+                android:layout_marginLeft="1dp"
+                android:layout_toRightOf="@id/etComment"
+                android:background="@android:color/transparent"
+                android:backgroundTint="@android:color/transparent"
+                android:text="Post"
+                android:textAllCaps="false"
+                android:textColor="#1778F2"
+                android:textSize="13sp" />
+        </RelativeLayout>
+
+
+    </RelativeLayout>
+</ScrollView>
