+++ app/src/main/res/layout/item_comment.xml
@@ -1,0 +1,47 @@
+<?xml version="1.0" encoding="utf-8"?>
+<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:layout_width="wrap_content"
+    android:layout_height="wrap_content">
+
+    <TextView
+        android:id="@+id/tvBody"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_alignParentEnd="true"
+        android:layout_marginStart="12dp"
+        android:layout_marginTop="10dp"
+        android:layout_marginEnd="15dp"
+        android:layout_toRightOf="@id/ivCommenterProfile"
+        android:minLines="1"
+        tools:text="heyhellojo Something random going in here helloooosdjflksdjflksjdflksjdflkjsdfklsjadflkajsdlkfjasljfdlasdjfl;asjdf;lajf" />
+
+    <TextView
+        android:id="@+id/tvTimestamp"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_below="@id/tvBody"
+        android:layout_marginStart="12dp"
+        android:layout_marginTop="5dp"
+        android:layout_toRightOf="@id/ivCommenterProfile"
+        android:textSize="11sp"
+        tools:text="1 day ago" />
+
+    <ImageView
+        android:id="@+id/ivCommenterProfile"
+        android:layout_width="40dp"
+        android:layout_height="40dp"
+        android:layout_alignParentStart="true"
+        android:layout_marginStart="15dp"
+        android:layout_marginTop="10dp"
+        app:srcCompat="@drawable/instagram_user_filled_24" />
+
+    <View
+        android:id="@+id/divider"
+        android:layout_width="match_parent"
+        android:layout_height="0.5dp"
+        android:layout_below="@id/tvTimestamp"
+        android:layout_marginTop="10dp"
+        android:background="@android:color/darker_gray" />
+</RelativeLayout>
