+++ app/src/main/res/layout/item_post.xml
@@ -150,8 +150,9 @@
         android:layout_alignStart="@id/tvDescription"
         android:layout_marginTop="5dp"
         android:textSize="14sp"
-        android:visibility="gone"
-        tools:text="View all 3 comments" />
+        android:visibility="visible"
+        tools:text="View all comments"
+        android:text="View all comments"/>
 
 
 </RelativeLayout>
